refactor: dedupe extractTrackingIssueUrl into scripts/lib
scripts/check-flaky-tracking-issue.mjs duplicated extractTrackingIssueUrl
from scripts/report-flaky-lane.mjs verbatim. Extract it to
scripts/lib/extract-tracking-issue-url.mjs, a side-effect-free module
both CLI scripts import — report-flaky-lane.mjs has argv-parsing /
process.exit side effects at module scope, so neither script imports
the other's CLI entry directly. Move the existing test coverage
(scripts/__tests__/report-flaky-lane.test.ts) to
extract-tracking-issue-url.test.ts, importing from the new module.

Ref zudolab/zudo-doc#2529
